Move Hasher and CredsGenerate to root package
