+++ authboss.go
@@ -39,6 +39,15 @@ func New() *Authboss {
 
 // Init authboss, modules, renderers
 func (a *Authboss) Init(modulesToLoad ...string) error {
+	// Create the hasher
+	// Have to do it in Init for backwards compatibility.
+	// If a user did not previously use defaults.SetCore() then they will
+	// suddenly start getting panics
+	// We also cannot use config.Defaults() so we can respect the user's BCryptCost
+	if a.Config.Core.Hasher == nil {
+		a.Config.Core.Hasher = NewBCryptHasher(a.Config.Modules.BCryptCost)
+	}
+
 	if len(modulesToLoad) == 0 {
 		modulesToLoad = RegisteredModules()
 	}
@@ -99,7 +108,7 @@ func VerifyPassword(user AuthableUser, password string) error {
 // in order to access the routes in protects. Requirements is a bit-set integer
 // to be able to easily combine requirements like so:
 //
-//   authboss.RequireFullAuth | authboss.Require2FA
+//	authboss.RequireFullAuth | authboss.Require2FA
 type MWRequirements int
 
 // MWRespondOnFailure tells authboss.Middleware how to respond to
@@ -156,7 +165,7 @@ func MountedMiddleware(ab *Authboss, mountPathed, redirectToLogin, forceFullAuth
 //
 // requirements are set by logical or'ing together requirements. eg:
 //
-//   authboss.RequireFullAuth | authboss.Require2FA
+//	authboss.RequireFullAuth | authboss.Require2FA
 //
 // failureResponse is how the middleware rejects the users that don't meet
 // the criteria. This should be chosen from the MWRespondOnFailure constants.
