+++ confirm/confirm_test.go
@@ -6,7 +6,6 @@ import (
 	"crypto/sha512"
 	"encoding/base64"
 	"errors"
-	"github.com/volatiletech/authboss/v3/defaults"
 	"net/http"
 	"net/http/httptest"
 	"testing"
@@ -72,8 +71,7 @@ func testSetup() *testHarness {
 
 	harness.ab.Config.Core.BodyReader = harness.bodyReader
 	harness.ab.Config.Core.Logger = mocks.Logger{}
-	harness.ab.Config.Core.Hasher = defaults.NewBCryptHasher(harness.ab.Modules.BCryptCost)
-	harness.ab.Config.Core.CredsGenerator = defaults.NewSha512CredsGenerator()
+	harness.ab.Config.Core.Hasher = mocks.Hasher{}
 	harness.ab.Config.Core.Mailer = harness.mailer
 	harness.ab.Config.Core.Redirector = harness.redirector
 	harness.ab.Config.Core.MailRenderer = harness.renderer
@@ -383,7 +381,7 @@ func TestMailURL(t *testing.T) {
 func TestGenerateRecoverCreds(t *testing.T) {
 	t.Parallel()
 
-	credsGenerator := defaults.NewSha512CredsGenerator()
+	credsGenerator := authboss.NewSha512CredsGenerator()
 
 	selector, verifier, token, err := credsGenerator.GenerateCreds()
 	if err != nil {
