+++ defaults/defaults.go
@@ -25,7 +25,5 @@ func SetCore(config *authboss.Config, readJSON, useUsername bool) {
 	config.Core.Redirector = NewRedirector(config.Core.ViewRenderer, authboss.FormValueRedirect)
 	config.Core.BodyReader = NewHTTPBodyReader(readJSON, useUsername)
 	config.Core.Mailer = NewLogMailer(os.Stdout)
-	config.Core.Hasher = NewBCryptHasher(config.Modules.BCryptCost)
-	config.Core.CredsGenerator = NewSha512CredsGenerator()
 	config.Core.Logger = logger
 }
