+++ go.mod
@@ -3,14 +3,19 @@ module github.com/volatiletech/authboss/v3
 go 1.19
 
 require (
+	github.com/friendsofgo/errors v0.9.2
+	github.com/pquerna/otp v1.4.0
+	golang.org/x/crypto v0.14.0
+	golang.org/x/oauth2 v0.6.0
+)
+
+require (
 	cloud.google.com/go v0.34.0 // indirect
 	github.com/boombuler/barcode v1.0.1 // indirect
 	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/friendsofgo/errors v0.9.2
 	github.com/golang/protobuf v1.5.3 // indirect
-	github.com/pquerna/otp v1.4.0
-	golang.org/x/crypto v0.7.0
-	golang.org/x/oauth2 v0.6.0
+	golang.org/x/net v0.10.0 // indirect
 	golang.org/x/xerrors v0.0.0-20220907171357-04be3eba64a2 // indirect
+	google.golang.org/appengine v1.6.7 // indirect
 	google.golang.org/protobuf v1.29.0 // indirect
 )
