+++ recover/recover_test.go
@@ -5,7 +5,6 @@ import (
 	"crypto/sha512"
 	"encoding/base64"
 	"errors"
-	"github.com/volatiletech/authboss/v3/defaults"
 	"net/http"
 	"net/http/httptest"
 	"strings"
@@ -86,8 +85,7 @@ func testSetup() *testHarness {
 
 	harness.ab.Config.Core.BodyReader = harness.bodyReader
 	harness.ab.Config.Core.Logger = mocks.Logger{}
-	harness.ab.Config.Core.Hasher = defaults.NewBCryptHasher(harness.ab.Config.Modules.BCryptCost)
-	harness.ab.Config.Core.CredsGenerator = defaults.NewSha512CredsGenerator()
+	harness.ab.Config.Core.Hasher = mocks.Hasher{}
 	harness.ab.Config.Core.Mailer = harness.mailer
 	harness.ab.Config.Core.Redirector = harness.redirector
 	harness.ab.Config.Core.MailRenderer = harness.renderer
@@ -473,7 +471,7 @@ func invalidCheck(t *testing.T, h *testHarness, w *httptest.ResponseRecorder) {
 func TestGenerateRecoverCreds(t *testing.T) {
 	t.Parallel()
 
-	credsGenerator := defaults.NewSha512CredsGenerator()
+	credsGenerator := authboss.NewSha512CredsGenerator()
 
 	selector, verifier, token, err := credsGenerator.GenerateCreds()
 	if err != nil {
